Update - Separated two similar but different sorting code implementations
There were two implementations of Quicksort.
The one synced with the display file was the RQS - 'randomized quicksort' implementation.
 The second, under the previous file name 'quicksort_LR', is the simple quicksort.
Both are linked to the main and presented separately
