+++ algorithms/quickSort.py
@@ -1,19 +1,23 @@
-import random
 from display import handleDrawing
-from random import randint
 
-def quickSort(array, left, right):
-    if left >= right:
-        return
-    index = left
-    random_index = randint(left, right)
-    array[right], array[random_index] = array[random_index], array[right]
+def quickSort(array, low, high):
+    if low < high:
+        p = partition(array, low, high)
+        quickSort(array, low, p)
+        quickSort(array, p + 1, high)
+
+def partition(array, low, high):
+    pivot = array[low]
+    i     = low - 1
+    j     = high + 1
     
-    for j in range(left, right):
-        handleDrawing(array, j, right, index, -1)
-        if array[j] < array[right]:
-            array[j], array[index] = array[index], array[j]
-            index += 1
-    array[index], array[right] = array[right], array[index]
-    quickSort(array, index + 1, right)
-    quickSort(array, left, index - 1)
+    while True:
+        i += 1
+        while array[i] < pivot: i += 1
+        
+        j -= 1
+        while array[j] > pivot: j -= 1
+        
+        handleDrawing(array, j, high, low, -1)
+        if i >= j: return j
+        array[i], array[j] = array[j], array[i]
