+++ algs.py
@@ -13,6 +13,7 @@ algorithmsDict = {'insertion'       : insertionSort,
                   'selection'       : selectionSort,
                   'merge'           : mergeSort,
                   'quick'           : quickSort,
+                  'randomized quick': randomizedquickSort,
                   'counting'        : countingSort,
                   'cocktail'        : cocktailSort,
                   'cycle'           : cycleSort,
